Converte Polegadas em Centímetros.

+++ ConvertePolegadasEmCentimetros.py
@@ -1,3 +1,3 @@
-valorEmPolegadas = int(input('Informe as polegadas: '))
+valorEmPolegadas = float(input('Informe as polegadas: '))
 centimetros = valorEmPolegadas*2.54
 print('Em Centimetros: ',centimetros)
